change namespace and assembly
