+++ App/Mixpanel.Net.WPF/MainWindow.xaml
@@ -7,6 +7,6 @@
         mc:Ignorable="d"
         Title="MainWindow" Height="350" Width="525">
     <Grid>
-        
+        <Button Content="Save" HorizontalAlignment="Center" VerticalAlignment="Center" Click="Button_Click" />
     </Grid>
 </Window>
