--- Library/Mixpanel.Net.Client.Universal/IO/ApplicationDataLocation.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mixpanel.Net.Client.Universal.IO
-{
-    public enum ApplicationDataLocation
-    {
-        Local,
-        Roaming,
-        Temp,
-        Install
-    }
-}
