--- Library/Mixpanel.Net.Client/Properties/AssemblyInfo.cs
@@ -1,30 +1,0 @@
-﻿using System.Resources;
-using System.Reflection;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
-
-// General Information about an assembly is controlled through the following 
-// set of attributes. Change these attribute values to modify the information
-// associated with an assembly.
-[assembly: AssemblyTitle("Mixpanel.Net.Client.SDK")]
-[assembly: AssemblyDescription("Integration Mixpanel track event for Windows Universal Apps and WPF.")]
-[assembly: AssemblyConfiguration("")]
-[assembly: AssemblyCompany("Pou")]
-[assembly: AssemblyProduct("Mixpanel.Net.Client.SDK")]
-[assembly: AssemblyCopyright("Copyright © Pou 2016")]
-[assembly: AssemblyTrademark("")]
-[assembly: AssemblyCulture("")]
-[assembly: NeutralResourcesLanguage("en")]
-
-// Version information for an assembly consists of the following four values:
-//
-//      Major Version
-//      Minor Version 
-//      Build Number
-//      Revision
-//
-// You can specify all the values or you can default the Build and Revision Numbers 
-// by using the '*' as shown below:
-// [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("1.0.0.1")]
-[assembly: AssemblyFileVersion("1.0.0.1")]
