--- Library/Mixpanel.Net.Client/ServiceModel/Parameter/EventParameter.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mixpanel.Net.Client.SDK.ServiceModel
-{
-    public class EventParameter: IParameter
-    {
-        public string ToQueryString()
-        {
-            return string.Empty;
-        }
-    }
-}
