--- Library/Mixpanel.Net.Client/ServiceModel/Parameter/IParameter.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mixpanel.Net.Client.SDK.ServiceModel
-{
-    interface IParameter
-    {
-        string ToQueryString();
-    }
-}
