--- Library/Mixpanel.Net.Client/Utility/CommonDefine.cs
@@ -1,19 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mixpanel.Net.Client.SDK
-{
-    public class CommonDefine
-    {
-        public const string TRACK_URI = "https://api.mixpanel.com/track";
-
-        public const int POSITION_TIMER_INTERVAL = 700;
-
-        public const string MIXPANEL_FOLDER_NAME = "MixpanelNetClientSDK";
-
-        public const string MIXPANEL_TRACK_FILE_NAME = "mixpanel_track_file.mtf";
-    }
-}
