+++ Library/MixpanelClientDotNet.UWP/IO/ApplicationDataLocation.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MixpanelDotNet.UWP.IO
+{
+    public enum ApplicationDataLocation
+    {
+        Local,
+        Roaming,
+        Temp,
+        Install
+    }
+}
