+++ Library/MixpanelClientDotNet.UWP/Properties/Mixpanel.Net.Client.Universal.rd.xml
