+++ Library/MixpanelClientDotNet/ServiceModel/Parameter/EventParameter.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MixpanelDotNet.ServiceModel
+{
+    public class EventParameter: IParameter
+    {
+        public string ToQueryString()
+        {
+            return string.Empty;
+        }
+    }
+}
