+++ Library/MixpanelClientDotNet/Utility/CommonDefine.cs
@@ -1,0 +1,19 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MixpanelDotNet
+{
+    public class CommonDefine
+    {
+        public const string TRACK_URI = "https://api.mixpanel.com/track";
+
+        public const int POSITION_TIMER_INTERVAL = 700;
+
+        public const string MIXPANEL_FOLDER_NAME = "MixpanelNetClientSDK";
+
+        public const string MIXPANEL_TRACK_FILE_NAME = "mixpanel_track_file.mtf";
+    }
+}
