+++ Library/MixpanelClientDotNet/Utility/INetworkHelper.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace MixpanelDotNet.Utility
+{
+    public interface INetworkHelper
+    {
+        bool IsNetworkAvailable { get; }
+    }
+}
