+++ Library/MixpanelClientDotNet/packages.config
