--- Mixpanel.Net.Client.WPF/Utility/NetworkHelper.cs
@@ -1,21 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Net.NetworkInformation;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Mixpanel.Net.Client.WPF.Utility
-{
-    public class NetworkHelper
-    {
-        public static bool IsNetworkAvailable
-        {
-            get
-            {
-                return NetworkInterface.GetIsNetworkAvailable();
-            }
-        }
-    }
-
-}
